requirement.txt
python-gitlab==3.2.0
gitpython==3.1.29